:hammer_and_wrench: refactoring - modularization
