+++ src/js/View/Templates.js
